% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the NEON Water Quality Shiny App}
\usage{
launch_app()
}
\description{
This function launches the Shiny application bundled with the package.
}
\examples{
# This example is not run to avoid hanging during checks
if (interactive()) {
  launch_app()
}
}
